ArgosApp/cloud_api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
ultralytics==8.3.225
pillow
fastapi==0.115.0
uvicorn[standard]==0.30.0
ultralytics==8.3.225
pillow
